move utils to separate file
